backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<1.0
PyYAML>=6.0,<7.0
langgraph>=0.2,<1.0
openai>=1.0,<2.0
httpx>=0.27,<1.0
beautifulsoup4>=4.12,<5.0
